requirements.txt
# ---------------------------------------------------- #
# install miniconda and create a new environment with:
# https://www.anaconda.com/docs/getting-started/miniconda/install#linux
# conda create -n fows python=3.10
# install the requirements with:
# conda activate fows
# pip install -r requirements.txt
# ---------------------------------------------------- #

# --------- python version --------- # 
# python==3.10.13 # NOTE: mediapipe requires python >= 3.9 and <= 3.12

# --------- pytorch --------- #
torch==2.1.2
torchvision==0.16.2

# --------- xception model --------- #
timm==1.0.3 # xception model

# --------- dataset preprocessing --------- #
mediapipe==0.10.20 # face detection and landmarks

# --------- image processing --------- #
pillow==10.0.1 # image processing
opencv-python==4.8.0.76 # image processing
grad-cam # gradcam visualization (https://github.com/jacobgil/pytorch-grad-cam)

# --------- other dependencies --------- #
scikit-learn==1.3.2 # metrics computation
numpy==1.26.4 # numerical operations
pandas==1.4.2 # data manipulation
matplotlib==3.8.2 # plotting 
